Fix changelog to compile breaking changes in a single release
We launched a breaking change in version v8.0.0 and v8.1.0. If someone was trying to figure out what was the breaking changes of the major version 8.x it would only read the v8.0.0 changelog. This commits removes v8.0.0 release and compiles the releases of v8.0.0, v8.0.1 and v8.1.0 in a single release, so this wouldnt be confusing.

Next time we will not do that and will launch a new major instead, avoiding this type of hack.