Add preferential sort order option to table
The Table component has a hardcoded preferential sort order set to 'ASC'
i.e. on a sortable column header first click, the table order is
automatically set to an ascending order.

However, in some applications, the most interesting numeric values in a
sortable list are the larger ones, and in the current context. These are
two clicks away instead of one.

To address this, add a `preferentialSortOrder`, defaulted to `ASC` --
the previous behaviour --, which can be used by the developer to set
the first sort order in a column sort click.