Fix SSR checking in Table
Change condition for checking SSR condition in Table to
`typeof document === 'undefined'` which is ReferenceError safe.

Current checking with `!document` yields a Reference Error in SSR, since
this global variable is not defined outside the browser--no DOM
references. Thus, a safer approach is to use `typeof` which evaluates
any symbol, even undeclared variables.