Update table height after rendering
On every first rendering of the Table, while using the
`dynamicRowHeight` option, we encounter a bug in which the table's
height is set to the default height set by its density, and then a
forced update (like a user interaction with the table) was needed
in order to make the table's height be the sum of the individual
rows' heights.

This happens because, when using `SimpleTable`'s
`calculateContainerHeight` on the render method, we failed to
notice that in order to correctly calculate the height needed to
fit the cell content into it, it was first necessary to let the
rendering occurr, and then measure each individual cell height.
Before that, the height needed for each cell is not even defined
to begin with.

Therefore, we must let a first rendering occurr with the default
table height ([density's corresponding line height] x [#lines]) and
then measure the needed line height for each cell. Since this
should happen after each table update (because content can change),
we place this in `componentDidUpdate` with a guard clause to prevent
excessive re-rendering.