Add page top indicator option to Pagination
Some pages have content that are big enough that the reference of which
page is being displayed is lost to the immediate view.

To fix this, add an option to the pagination component in which a page
indicator is shown above the content in such a way that the partial
nature of a search result is easier to perceive.