Add id to numeric stepper input (#1454)
#### What is the purpose of this pull request?

This is the second of two PRs to solve a web accessibility issue for the
NumberStepper input and associated label. The first PR is here:
https://github.com/vtex-apps/product-quantity/pull/68

This PR adds the id of "vtex-product-quantity-input" to the input
element, which is labelled by the label element created in the first PR
link above.

#### How to test it?

Check the HTML in developer console to see the changes reflected.


[Workspace](https://accessibilitytesting--colprofr.myvtex.com/elmex-kids-junior-toothpaste-61044604/p)

#### Screenshots or example usage:

<img width="1726" height="912" alt="Screenshot 2025-08-19 at 09 54 56"
src="https://github.com/user-attachments/assets/469cd6b4-046f-489f-906f-6474627f9962"
/>

#### Screenshots or example usage

#### Types of changes

- [x] Bug fix (a non-breaking change which fixes an issue)
- [ ] New feature (a non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to change)
- [ ] Requires change to documentation, which has been updated
accordingly.